requirements.txt
# Dark Factory Engine Dependencies
pyyaml>=6.0
pathlib2>=2.3.7; python_version < '3.4'